Add Custom Exceptions
Added ```MagentoError``` custom exception class for bad responses from the API
* Properly parses the error message to include error parameters
*  Base exception class for more specific exceptions, like ```AuthenticationError``
    - Not sure any other subclasses are needed tbh
* Updated Client to use new AuthenticationError
